kuksa-databroker/kuksa-ditto/requirements.txt
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
cmd2==1.5.0
colorama==0.4.6
eclipse-zenoh==1.8.0
grpcio==1.78.0
grpcio-tools==1.68.0
idna==3.11
jsonpath-ng==1.8.0
kuksa_client==0.5.1
protobuf==5.29.6
Pygments==2.19.2
pyperclip==1.11.0
requests==2.32.5
setuptools==82.0.1
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
